+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^6.4.4",
     "@mui/material": "^6.4.4",
     "@radix-ui/react-accordion": "^1.2.2",
     "@radix-ui/react-dialog": "^1.1.3",
@@ -28,9 +29,12 @@
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18.3.1",
+    "react-html-parser": "^2.0.2",
     "react-quill": "^2.0.0",
     "react-redux": "^9.1.2",
     "react-router": "^7.0.2",
+    "react-router-dom": "^7.1.5",
+    "react-tag-input-component": "^2.0.2",
     "react-toastify": "^10.0.6",
     "recharts": "^2.15.0",
     "tailwind-merge": "^2.5.5",
@@ -49,6 +53,6 @@
     "globals": "^15.9.0",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.12",
-    "vite": "^5.4.1"
+    "vite": "^5.4.14"
   }
 }
